replace generate, replace tests
