+++ apps/extension/src/state/seed-phrase/generate.ts
@@ -4,17 +4,8 @@ import {
   generateSeedPhrase,
   SeedPhraseLength,
   ValidationField,
+  generateValidationFields,
 } from '@penumbra-zone/crypto-web/src/mnemonic';
-
-export const generateValidationFields = (
-  seedPhrase: string[],
-  amount: number,
-): ValidationField[] => {
-  const allWords: ValidationField[] = seedPhrase.map((word, index) => ({ word, index }));
-  const shuffleWords = allWords.sort(() => 0.5 - Math.random());
-  const pickWords = shuffleWords.slice(0, amount);
-  return pickWords.sort((a, b) => a.index - b.index);
-};
 
 export interface GenerateFields {
   phrase: string[];
